AI Invoice Manager/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
google-cloud-vision==3.4.5
google-generativeai==0.3.2
sqlalchemy==2.0.23
sqlite3
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2